backend/python-worker/requirements-test.txt
# ==========================================
# Testing Dependencies
# ==========================================
# Python 3.11+ required
# For comprehensive testing of EC2 file processing pipeline

# ==========================================
# Core Testing Frameworks
# ==========================================
pytest==8.3.2
pytest-cov==5.0.0
pytest-xdist==3.6.1  # Parallel test execution
pytest-timeout==2.3.1  # Test timeout handling
pytest-mock==3.14.0  # Enhanced mocking capabilities
pytest-asyncio==0.24.0  # Async test support

# ==========================================
# AWS Service Mocking
# ==========================================
moto[all]==5.0.13  # Mock AWS services (S3, SQS, OpenSearch)
boto3-stubs[s3,sqs,opensearch]==1.34.144  # Type hints for boto3

# ==========================================
# Test Data Generation
# ==========================================
faker==28.1.0  # Generate realistic test data
hypothesis==6.111.2  # Property-based testing
factory-boy==3.3.1  # Test fixture factories

# ==========================================
# Performance Testing
# ==========================================
locust==2.31.5  # Load testing framework
pytest-benchmark==4.0.0  # Performance benchmarking

# ==========================================
# Code Quality
# ==========================================
pylint==3.2.7  # Static code analysis
black==24.8.0  # Code formatting
isort==5.13.2  # Import sorting
mypy==1.11.2  # Type checking
flake8==7.1.1  # Style guide enforcement

# ==========================================
# Coverage and Reporting
# ==========================================
coverage[toml]==7.6.1  # Code coverage
pytest-html==4.1.1  # HTML test reports
allure-pytest==2.13.5  # Advanced test reporting

# ==========================================
# Mocking and Fixtures
# ==========================================
responses==0.25.3  # Mock HTTP requests
freezegun==1.5.1  # Mock datetime
pillow==10.4.0  # Image generation for tests

# ==========================================
# Document Testing
# ==========================================
# For creating test documents
pypdf==4.3.1  # PDF creation
python-docx==1.1.2  # Word document creation
openpyxl==3.1.5  # Excel creation
python-pptx==1.0.2  # PowerPoint creation

# ==========================================
# Utilities
# ==========================================
pytest-env==1.1.5  # Environment variable management
pytest-watch==4.2.0  # Auto-run tests on file changes
